pyproject.toml
[tool.poetry]
name = "fastapi-p21"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
jinja2 = "^3.1.4"
sqladmin = "^0.19.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
sqlalchemy-file = "^0.6.0"
pillow = "^10.4.0"
fastapi-storages = "^0.3.0"
itsdangerous = "^2.2.0"
asyncpg = "^0.29.0"
greenlet = "^3.1.1"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
